is there enough components (??)) PLEASE REVIEW :3
