+++ frontend/src/components/GamePage/TabsComponent.vue
@@ -1,0 +1,32 @@
+<template>
+    <div id="themes">
+      <Tabs>
+        <TabList class="tabs">
+          <Tab v-for="tab in tabs" :value="tab.value" :key="tab">{{ tab.name }}</Tab>
+        </TabList>
+        <TabPanels>
+        </TabPanels>
+      </Tabs>
+    </div>
+</template>
+
+<script setup lang="ts">
+import Tabs from 'primevue/tabs'
+import TabList from 'primevue/tablist'
+import Tab from 'primevue/tab'
+import TabPanels from 'primevue/tabpanels'
+
+const tabs = [
+  { name: 'Randomized', value: '0' },
+  { name: 'Realistic', value: '1' },
+  { name: 'Anime', value: '2' },
+  { name: 'Photography', value: '3' },
+  { name: 'Still Life', value: '4' }
+]
+</script>
+
+<style scoped>
+.tabs {
+  align-self: center;
+}
+</style>
